@@ -1,6 +1,6 @@
 {
   "name": "studioyosepra-ui",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
@@ -9,29 +9,35 @@
     "deploy": "gh-pages -d dist"
   },
   "dependencies": {
-    "@contentful/rich-text-react-renderer": "^15.2.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
-    "bootstrap": "^4.6.0",
-    "classnames": "^2.3.1",
-    "contentful": "^8.5.3",
-    "react": "^17.0.0",
-    "react-bootstrap": "^1.6.1",
-    "react-dom": "^17.0.0"
+    "@contentful/rich-text-react-renderer": "^15.22.11",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-brands-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@reduxjs/toolkit": "^2.5.0",
+    "axios": "^1.7.9",
+    "bootstrap": "^5.3.3",
+    "classnames": "^2.5.1",
+    "contentful": "^11.3.4",
+    "date-fns": "^4.1.0",
+    "formik": "^2.4.6",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.7",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^6.28.1"
   },
   "devDependencies": {
-    "@vitejs/plugin-react-refresh": "^1.3.1",
-    "eslint": "^8.2.0",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "eslint": "^8.57.1",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.3",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "faker": "^5.5.3",
-    "gh-pages": "^3.2.3",
-    "sass": "^1.38.2",
-    "vite": "^2.5.2"
+    "gh-pages": "^6.2.0",
+    "sass": "^1.83.0",
+    "vite": "^2.9.18"
   }
 }
